Refactor: Split large files and add author reference
Split files larger than 350 lines:
- djangovue/settings.py into settings.py and utils.py
- backend/tests.py into a tests package containing test_views.py, test_security.py, and test_settings.py

Added the required author reference 'Mike Borozdin (mikebz@)' at the top of all .py and .js files.
