+++ backend/__init__.py
@@ -1,1 +1,2 @@
+# Author: Mike Borozdin (mikebz@)
 """Backend Django application package."""
