+++ backend/admin.py
@@ -1,1 +1,2 @@
+# Author: Mike Borozdin (mikebz@)
 """Django admin registrations for backend models."""
