+++ backend/models.py
@@ -1,1 +1,2 @@
+# Author: Mike Borozdin (mikebz@)
 """Database models for the backend Django app."""
