+++ backend/tests/__init__.py
@@ -1,0 +1,1 @@
+# Author: Mike Borozdin (mikebz@)
