+++ backend/urls.py
@@ -1,3 +1,4 @@
+# Author: Mike Borozdin (mikebz@)
 """URL patterns for backend views."""
 
 from django.urls import URLPattern, path
