+++ djangovue/__init__.py
@@ -1,1 +1,2 @@
+# Author: Mike Borozdin (mikebz@)
 """Django project package for djangovue."""
