+++ djangovue/asgi.py
@@ -1,3 +1,4 @@
+# Author: Mike Borozdin (mikebz@)
 """ASGI config for djangovue project.
 
 It exposes the ASGI callable as a module-level variable named ``application``.
