+++ djangovue/wsgi.py
@@ -1,3 +1,4 @@
+# Author: Mike Borozdin (mikebz@)
 """WSGI config for djangovue project.
 
 It exposes the WSGI callable as a module-level variable named ``application``.
