+++ frontend/js/main.js
@@ -1,3 +1,4 @@
+// Author: Mike Borozdin (mikebz@)
 import { createApp } from 'vue'
 import App from './App.vue'
