+++ manage.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+# Author: Mike Borozdin (mikebz@)
 """Django command-line utility for administrative tasks.
 
 The implementation lives in djangovue/cli.py so that it can also be installed
