+++ vite.config.js
@@ -1,3 +1,4 @@
+// Author: Mike Borozdin (mikebz@)
 import { defineConfig } from 'vite'
 import vue from '@vitejs/plugin-vue'
